[I25-412] feat : AutoSlidingBusinessCard를 서버 컴포넌트로 변경하여 내부에서 데이터 페칭
